Refactor signatures. Track if precontracts come from users or not
