+++ creusot/src/backend/logic/vcgen.rs
@@ -13,7 +13,8 @@ use crate::{
     },
     contracts_items::get_builtin,
     pearlite::{super_visit_term, Literal, Pattern, PointerKind, Term, TermVisitor},
-    util::{self, pre_sig_of},
+    signature::pre_sig_of,
+    util::{self},
 };
 use rustc_hir::def_id::DefId;
 use rustc_middle::ty::{EarlyBinder, GenericArgsRef, ParamEnv, Ty, TyKind};
@@ -426,12 +427,12 @@ impl<'a, 'tcx> VCGen<'a, 'tcx> {
             // TODO: lol
             TermKind::Absurd => todo!("absrd"),
 
-            TermKind::Precondition {..} => Err(VCError::Closure(t.span)),
-            TermKind::Postcondition {..} => Err(VCError::Closure(t.span)),
+            TermKind::Precondition { .. } => Err(VCError::Closure(t.span)),
+            TermKind::Postcondition { .. } => Err(VCError::Closure(t.span)),
             TermKind::Old { .. } => Err(VCError::Old(t.span)),
             TermKind::Closure { .. } => Err(VCError::Closure(t.span)),
             TermKind::Reborrow { .. } => Err(VCError::Reborrow(t.span)),
-            TermKind::Borrow { inner } => todo!(),
+            TermKind::Borrow { inner: _ } => todo!(),
         }
     }
     fn build_pattern<A>(
