+++ creusot/src/backend/signature.rs
@@ -6,10 +6,7 @@ use why3::{
 };
 
 use crate::{
-    backend,
-    contracts_items::{should_replace_trigger, why3_attrs},
-    translation::specification::PreContract,
-    util::{ident_of, item_name, AnonymousParamName, PreSignature},
+    backend, contracts_items::{should_replace_trigger, why3_attrs}, signature::PreSignature, translation::specification::PreContract, util::{ident_of, item_name, AnonymousParamName}
 };
 
 use super::{logic::function_call, term::lower_pure, Namer, Why3Generator};
